Add parallel flag as a temporary fix.
This is necessary because the creation reduction context expects all
processors inside the MPI cluster to participate. If
`pll_create_partition` is called from a single processor, the others
will never call `new_reduction_context` and the process will hang.

If the parallel flag is set to 0, we will not create a reduction context
and expect all calculations to be carried out on a singular process.